refactor: 공간 예약 완료 스크린 Stack 이동
